fix: resolve all 433 open CodeQL code-scanning alerts
121 code fixes + 312 documented dismissals (false positives and
intentional teaching code), covering every open alert on main.

Code fixes:
- Path.Combine → Path.Join (66): the cs/path-combine query flags every
  Combine call; Join is byte-identical for these inputs.
- readonly fields (24), dead stores removed (15), using/Dispose (13),
  redundant casts removed (14), LINQ Where/Select conversions (8),
  boolean simplifications, rethrow, float cast for an int-multiplication
  overflow (error-severity), type-identity asserts replacing fragile
  FullName string comparisons, comments in expected-cancellation catches.

Dismissals (each with a reason comment on the alert):
- 173 obj/ source-generator outputs (classifications: generated).
- 139 intentional teaching code or verified false positives: P/Invoke
  & GC lessons, 'throw ex' anti-pattern demos marked with #pragma
  CA2200, foreach desugaring, mutation-during-enumeration, boxing
  copy-semantics, hand-rolled Where, double-Dispose idempotence,
  ALC root-dropping, anonymous-type null-cast inference, escaped
  disposables (WebApplicationFactory/HttpResponseMessage contracts).

Verified: LearnCSharp.CI.slnx Debug+Release build 0 warnings / 0 errors;
dotnet format --verify-no-changes clean; all 271 LearnCSharp topics run
green with asserts active; 236/236 tests pass against a real Docker
stack (postgres, redis, rabbitmq, kafka, mailpit, keycloak); all
pre-commit hooks pass.

+++ src/LearnCSharp/Stage12_PerformanceLine/Section04_SimdAndVectorization/SimdConcept.cs
@@ -61,7 +61,7 @@ internal static class SimdConcept
             scalar += data[i];
 
         // conceptual parallel add (full Vector demos in later topics)
-        float check = data.Length * (data.Length - 1) / 2f;
+        float check = (float)data.Length * (data.Length - 1) / 2f;
         Debug.Assert(Math.Abs(scalar - check) < 0.1f);
         Console.WriteLine($"  scalar sum of 0..1023 = {scalar}");
         Console.WriteLine("  Use SIMD last: after profile + after reducing allocations.");
